MNT: Reorganize callbacks modules into callbacks package.
