+++ examples/plotting_2d.py
@@ -1,6 +1,6 @@
 from bluesky.examples import *
 from bluesky.broker_examples import *
-from bluesky.broker_callbacks import LiveImage
+from bluesky.callbacks.broker import LiveImage
 from bluesky.tests.utils import setup_test_run_engine
 from matplotlib import pyplot as plt
 from xray_vision.backend.mpl.cross_section_2d import CrossSection
